+++ src/include/math_utils.h
@@ -1,0 +1,9 @@
+#ifndef CUHKSZ_MATH
+#define CUHKSZ_MATH
+
+#include "math_utils/BigNum.h"
+#include "math_utils/Fraction.h"
+#include "math_utils/Matrix.h"
+#include "math_utils/math_functions.h"
+
+#endif  // CUHKSZ_MATH
