--- src/include/mathematics.h
@@ -1,9 +1,0 @@
-#ifndef CUHKSZ_MATH
-#define CUHKSZ_MATH
-
-#include "math/BigNum.h"
-#include "math/Fraction.h"
-#include "math/Matrix.h"
-#include "math/math_functions.h"
-
-#endif  // CUHKSZ_MATH
